Break auth server out into separate binary
